Package to check security advisories
This package just adds conflicts with compose packages that have known
vulnerabilities in, so will highlight packages that are being used that
contain vulnerabilities.

This should be updated regularly to update the signature with the latest
discovered vulnerabilities.

Should this packages conflict with another package, then you should look
at patching that package to a secured version.